+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -2,11 +2,11 @@ object Dependencies {
     const val desugar = "com.android.tools:desugar_jdk_libs:1.1.5"
 
     object ComposeMaterialDialogs {
-        const val version = "0.6.2"
+        const val version = "0.6.3"
 
-        const val core = "io.github.vanpra.compose-material-dialogs:core:$version"
-        const val datetime = "io.github.vanpra.compose-material-dialogs:datetime:$version"
-        const val color = "io.github.vanpra.compose-material-dialogs:color:$version"
+        const val core = "ca.gosyer:compose-material-dialogs-core:$version"
+        const val datetime = "ca.gosyer:compose-material-dialogs-datetime:$version"
+        const val color = "ca.gosyer:compose-material-dialogs-color:$version"
     }
 
     object Ktlint {