+++ core/src/desktopMain/kotlin/com/vanpra/composematerialdialogs/DesktopUtils.kt
@@ -1,20 +1,35 @@
 package com.vanpra.composematerialdialogs
 
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.platform.LocalWindowInfo
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.DpSize
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogState
+import androidx.compose.ui.window.DialogWindowScope
 import androidx.compose.ui.window.WindowPosition
+import androidx.compose.ui.window.rememberDialogState
 
+private fun DpSize.toScreenConfiguration(): ScreenConfiguration {
+    return ScreenConfiguration(
+        width.value.toInt(),
+        height.value.toInt()
+    )
+}
 
 @Composable
 internal actual fun rememberScreenConfiguration(): ScreenConfiguration {
-    return remember {
-        ScreenConfiguration(
-            screenWidthDp = 600,
-            screenHeightDp = 400
-        )
-    }
+    return LocalScreenConfiguration.current
 }
 
 @Composable
@@ -27,6 +42,8 @@ internal actual fun isLargeDevice(): Boolean {
     return true
 }
 
+internal val LocalScreenConfiguration = compositionLocalOf<ScreenConfiguration>{ throw IllegalStateException("Unused") }
+
 @Composable
 internal actual fun DialogBox(
     onDismissRequest: () -> Unit,
@@ -34,9 +51,20 @@ internal actual fun DialogBox(
     content: @Composable () -> Unit
 ) = Dialog(
     onCloseRequest = onDismissRequest,
-    state = DialogState(position = properties.position.toWindowPosition(), size = properties.size)
+    state = rememberDialogState(position = properties.position.toWindowPosition(), size = properties.size),
+    title = properties.title,
+    icon = properties.icon,
+    resizable = properties.resizable
 ) {
-    content()
+    BoxWithConstraints {
+        CompositionLocalProvider(
+            LocalScreenConfiguration provides ScreenConfiguration(
+                maxWidth.value.toInt(),
+                maxHeight.value.toInt()
+            ),
+            content = content
+        )
+    }
 }
 
 private fun DesktopWindowPosition.toWindowPosition(): WindowPosition {
@@ -45,4 +73,25 @@ private fun DesktopWindowPosition.toWindowPosition(): WindowPosition {
         is DesktopWindowPosition.Absolute -> WindowPosition(x = x, y = y)
         is DesktopWindowPosition.Aligned -> WindowPosition(alignment)
     }
+}
+
+@Composable
+internal actual fun defaultDialogShape() = RectangleShape
+
+@Composable
+internal actual fun ScreenConfiguration.getMaxHeight(): Dp {
+    return screenHeightDp.dp
+}
+
+@Composable
+internal actual fun ScreenConfiguration.getPadding(maxWidth: Dp): Dp {
+    return 0.dp
+}
+
+internal actual fun Modifier.dialogHeight(): Modifier = Modifier.fillMaxHeight()
+
+internal actual fun Modifier.dialogMaxSize(maxHeight: Dp): Modifier = this
+
+internal actual fun getLayoutHeight(maxHeightPx: Int, layoutHeight: Int): Int {
+    return maxHeightPx
 }