+++ desktop/src/main/kotlin/com/vanpra/composematerialdialogdemos/Strings.kt
@@ -1,0 +1,11 @@
+package com.vanpra.composematerialdialogdemos
+
+object Strings {
+    const val location_dialog_message = "Let us help apps determine location. This means sending\n" +
+        "    anonymous location data to us, even when no apps are running"
+    const val location_dialog_title = "Use Location Services?"
+    const val ringtone_dialog_title = "Phone Ringtone"
+    const val labels_dialog_title = "Label as:"
+    const val backup_dialog_title = "Set backup account"
+    const val input_dialog_title = "Enter you name:"
+}